+++ CHANGES.rst
@@ -10,12 +10,11 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 4.4.5 (unreleased)
 ------------------
 
-- support the ``attrs`` predefined template variable again (as
+- Support the ``attrs`` predefined template variable again (as
   far as ``chameleon`` allows it)
   (`#860 <https://github.com/zopefoundation/Zope/issues/860>`_).
 
-- in the "Zope Book" remove ``CONTEXTS`` from the list of predefined
-  template variables (because it was never implemented).
+- Improve documentation of ``CONTEXTS`` in the "Zope Book".
   
 - Update dependencies to the latest releases that still support Python 2.